Refactor useKeypress to avoid edge cases when keypress handlers were added and removed
