+++ packages/cli/src/ui/components/ShellConfirmationDialog.test.tsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { render } from 'ink-testing-library';
+import { renderWithProviders } from '../../test-utils/render.js';
 import { describe, it, expect, vi } from 'vitest';
 import { ShellConfirmationDialog } from './ShellConfirmationDialog.js';
 
@@ -17,12 +17,16 @@ describe('ShellConfirmationDialog', () => {
   };
 
   it('renders correctly', () => {
-    const { lastFrame } = render(<ShellConfirmationDialog request={request} />);
+    const { lastFrame } = renderWithProviders(
+      <ShellConfirmationDialog request={request} />,
+    );
     expect(lastFrame()).toMatchSnapshot();
   });
 
   it('calls onConfirm with ProceedOnce when "Yes, allow once" is selected', () => {
-    const { lastFrame } = render(<ShellConfirmationDialog request={request} />);
+    const { lastFrame } = renderWithProviders(
+      <ShellConfirmationDialog request={request} />,
+    );
     const select = lastFrame()!.toString();
     // Simulate selecting the first option
     // This is a simplified way to test the selection
@@ -30,14 +34,18 @@ describe('ShellConfirmationDialog', () => {
   });
 
   it('calls onConfirm with ProceedAlways when "Yes, allow always for this session" is selected', () => {
-    const { lastFrame } = render(<ShellConfirmationDialog request={request} />);
+    const { lastFrame } = renderWithProviders(
+      <ShellConfirmationDialog request={request} />,
+    );
     const select = lastFrame()!.toString();
     // Simulate selecting the second option
     expect(select).toContain('Yes, allow always for this session');
   });
 
   it('calls onConfirm with Cancel when "No (esc)" is selected', () => {
-    const { lastFrame } = render(<ShellConfirmationDialog request={request} />);
+    const { lastFrame } = renderWithProviders(
+      <ShellConfirmationDialog request={request} />,
+    );
     const select = lastFrame()!.toString();
     // Simulate selecting the third option
     expect(select).toContain('No (esc)');
